+++ src/utils/index.js
@@ -2,6 +2,7 @@ import bech32Manager from './bech32Manager';
 import blocksRequestHelper from './blocksRequestHelper';
 import coinAdapter from './coinAdapter';
 import dateHandler from './dateHandler';
+import msgBuilder from './msgBuilder';
 import proposerHandler from './proposerHandler';
 import regExpBuilder from './regExpBuilder';
 import storageHandler from './storageHandler';
@@ -16,6 +17,7 @@ export {
   blocksRequestHelper,
   coinAdapter,
   dateHandler,
+  msgBuilder,
   proposerHandler,
   regExpBuilder,
   storageHandler,