+++ CHANGELOG.md
@@ -5,14 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [4.8.1] - 2023-09-22
+## [alfa] 4.8.2 - 2023-09-22
+
+### Added
+
+- Add assets section:
+  - CW20 management.
+
+## [alfa] 4.8.1 - 2023-09-22
 
 ### Added
 
 - Add assets section:
   - CW20 detail.
 
-## [develop] 4.8.0 - 2023-09-21
+## [alfa] 4.8.0 - 2023-09-21
 
 ### Added